requirements.txt
arrow==0.10.0
Babel==2.3.4
click==6.6
Flask==0.12
Flask-Login==0.4.0
Flask-PyMongo==0.4.1
Flask-Table==0.2.13
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8.1
MarkupSafe==0.23
pymongo==3.4.0
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
Werkzeug==0.11.15
WTForms==2.1
biopython-1.70
numpy-1.13.3
